make validate functions raise exceptions by default
