--- tests/test_validate.py
@@ -1,66 +1,0 @@
-from fastavro.validation import validate
-
-
-def test_validator_numeric():
-    for datum, schema in [
-        (1, 'int'),
-        (1, 'long'),
-        (1.0, 'float'),
-        (1.0, 'double'),
-        (1, 'float'),
-        (1, 'double'),
-    ]:
-        assert validate(datum, schema)
-
-    for datum, schema in [
-        (1.0, 'int'),
-        (1.0, 'long'),
-        ("1.0", 'float'),
-        ("1.0", 'double'),
-        ("1", 'float'),
-        ("1", 'double'),
-    ]:
-        assert not validate(datum, schema)
-    # and plenty more to add I suppose
-
-
-def test_validator_numeric_numpy():
-    import numpy as np
-    np_ints = [
-        np.int_,
-        np.intc,
-        np.intp,
-        np.int8,
-        np.int16,
-        np.int32,
-        np.int64,
-        np.uint8,
-        np.uint16,
-        np.uint32,
-        np.uint64,
-    ]
-
-    np_floats = [
-        np.float_,
-        np.float16,
-        np.float32,
-        np.float64,
-    ]
-
-    schema_ints = ['int', 'long']
-
-    schema_floats = ['float', 'double']
-
-    # all these should work
-    for nptype, schema in zip(np_ints, schema_ints):
-        assert validate(nptype(1), schema)
-
-    for nptype, schema in zip(np_ints, schema_floats):
-        assert validate(nptype(1), schema)
-
-    for nptype, schema in zip(np_floats, schema_floats):
-        assert validate(nptype(1), schema)
-
-    # these shouldn't work
-    for nptype, schema in zip(np_floats, schema_ints):
-        assert not validate(nptype(1), schema)
